# octbase-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octbase API
  version: "0.4.0"
  description: |
    REST API for Octbase — a focused project-management tool (tasks, boards,
    releases, sprints, wiki pages, SCM integration) for a single, invitation-only
    team.

    **Authentication:** All routes require a JWT **Bearer** access token in the
    `Authorization` header, except the public auth/invitation routes, the
    HMAC-authenticated webhook receivers, and the unauthenticated ops endpoints
    (`/api/v1/health`, `/api/v1/version`, `/api/v1/meta/enums`, `/api/v1/config`, `/metrics`,
    `/openapi.yaml`, `/docs`). Obtain an access token from `POST /api/v1/auth/login`;
    it is returned as `{ "accessToken", "expiresIn" }`. A long-lived refresh token is
    set as an HttpOnly cookie and rotated on every `POST /api/v1/auth/refresh`.

    **Authorization:** Two layers, both enforced server-side. A global role
    (`SUPER_ADMIN`, `ADMIN`, `USER`, `GUEST`) and a per-project membership role
    (`PROJECT_OWNER`, `PROJECT_ADMIN`, `PROJECT_MEMBER`, `PROJECT_VIEWER`). The
    effective project permission matrix is available at
    `GET /api/v1/projects/{projectId}/permissions`.

    **Error format:** All errors return
    `{ "code": "ERROR_CODE", "message": "...", "messageKey": "...", "details": {...} }`
    (`messageKey`/`details` optional).

    **Pagination:** List endpoints accept `?page=0&size=20` (zero-indexed pages,
    default size 20).

...
  contact:
    name: Octbase
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page: { name: page, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
    size: { name: size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200, default: 20 } }
...
  responses:
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Conflict / version mismatch, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Business-rule violation (e.g. LAST_OWNER, RELEASE_HAS_OPEN_TASKS), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        messageKey: { type: string }
        details: { type: object, additionalProperties: true }
...
    MfaEnrollment:
      type: object
      properties:
        otpauthUrl: { type: string, description: otpauth:// URI for client-side QR rendering }
        secret: { type: string, description: Manual-entry fallback for the same secret }
    MfaRecoveryCodes:
      type: object
      properties:
        recoveryCodes:
          type: array
          items: { type: string }
          description: Shown once at generation time; only hashes are stored thereafter.
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        kind: { type: string }
        projectId: { type: string, format: uuid }
        taskId: { type: string, format: uuid }
        pageId: { type: string, format: uuid }
        message: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    NotificationPreference:
      type: object
      properties:
        userId: { type: string, format: uuid }
        kind: { type: string }
        inApp: { type: boolean }
        email: { type: boolean }
paths:
...
  /api/v1/users/me/notifications:
    get:
      tags: [Notifications]
      summary: List my notifications
      operationId: listNotifications
      parameters: [ { $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/size' } ]
      responses:
        '200': { description: Notifications, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Notification' } } } } }
  /api/v1/users/me/notifications/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    patch:
      tags: [Notifications]
      summary: Mark a notification read/unread
      operationId: updateNotification
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { isRead: { type: boolean } } } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Notification' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/users/me/notification-preferences:
    get:
      tags: [Notifications]
      summary: Get my per-kind notification preferences
      operationId: getNotificationPreferences
      responses:
        '200': { description: Preferences, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/NotificationPreference' } } } } }
    patch:
      tags: [Notifications]
      summary: Update my notification preferences
      operationId: updateNotificationPreferences
      requestBody: { required: true, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/NotificationPreference' } } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/NotificationPreference' } } } } }
...
  /api/v1/users/me/mfa/enroll:
    post:
      tags: [MFA]
      summary: Begin MFA enrollment (generates a pending TOTP secret)
      description: >
        Requires the account password for re-authentication when called with a
        normal access token, so a stolen access token alone cannot enable MFA.
        The forced-enrollment-at-login flow (scoped enrollment token) is exempt.
      operationId: enrollMfa
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { password: { type: string, format: password } } } } } }
      responses:
        '200': { description: Enrollment started, content: { application/json: { schema: { $ref: '#/components/schemas/MfaEnrollment' } } } }
        '401': { description: Password re-authentication required (stable code REAUTH_REQUIRED) }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/users/me/mfa/confirm:
    post:
      tags: [MFA]
      summary: Confirm MFA enrollment with a TOTP code, activating MFA
      operationId: confirmMfa
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [code], properties: { code: { type: string } } } } } }
      responses:
        '200': { description: MFA enabled, content: { application/json: { schema: { $ref: '#/components/schemas/MfaRecoveryCodes' } } } }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/users/me/mfa/disable:
    post:
      tags: [MFA]
      summary: Disable MFA (requires the current password or a valid code)
      operationId: disableMfa
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { password: { type: string }, code: { type: string } } } } } }
      responses:
        '204': { description: MFA disabled }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/users/me/mfa/recovery-codes/regenerate:
    post:
      tags: [MFA]
      summary: Regenerate recovery codes (requires the current password or a valid code)
      operationId: regenerateMfaRecoveryCodes
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { password: { type: string }, code: { type: string } } } } } }
      responses:
        '200': { description: New recovery codes, content: { application/json: { schema: { $ref: '#/components/schemas/MfaRecoveryCodes' } } } }
        '422': { $ref: '#/components/responses/Unprocessable' }

  # ───────────────────────── Projects ─────────────────────────
...